requirements.txt
qrcode==7.4.2
requests==2.32.3
watchdog==4.0.1
moviepy==1.0.3
numpy==1.26.2
Pillow==10.1.0
ffpyplayer==4.5.1